requirements.txt
google-cloud-bigquery
google-cloud-storage
requests
python-dotenv
boto3
requests-aws4auth
openpyxl
